prune obsolete type conversions
git-svn-id: https://svn.broadinstitute.org/gaag/pilon/trunk@3741 333ed61c-d966-4eaf-8115-f610d1735cec

+++ src/main/scala/org/broadinstitute/pilon/GenomeRegion.scala
@@ -133,10 +133,10 @@ class GenomeRegion(val contig: ReferenceSequence, start: Int, stop: Int)
       val r = refBase(i + start)
       //val minDepth = meanCoverage * 25 / 100
       val loc = locus(i)
-      coverage(i) = n.toShort
-      badCoverage(i) = pu.badPair.toShort
-      physCoverage(i) = pu.physCov.toShort
-      insertSize(i) = pu.insertSize.toShort
+      coverage(i) = n.toInt
+      badCoverage(i) = pu.badPair
+      physCoverage(i) = pu.physCov
+      insertSize(i) = pu.insertSize
       weightedQual(i) = pu.weightedQual.toByte
       weightedMq(i) = pu.weightedMq.toByte
       clips(i) = pu.clips.toShort
